Weapons can now shoot towards enemies
- [Added] laser cannon bullet prefab
- [Modified] Laser Cannon ammo script to fire it straight towards a target

+++ Assets/_TowerDefense3D/Scripts/ScriptableObjects/Placeables/RangeWeaponAttributes.cs
@@ -6,7 +6,7 @@ namespace TowerDefense3D
     public class RangeWeaponAttributes : WeaponItemAttributes
     {
         public AmmoAttributes ammo;
-        public int fireRate;
+        public float fireRate;
         
         [Header("Rotation")]
         public float trackTargetSpeed;
